Moved some functions to new files + some comments
