+++ API/http.go
@@ -2,15 +2,12 @@ package main
 
 import (
 	"database/sql"
-	"encoding/json"
 	"fmt"
 	_ "github.com/go-sql-driver/mysql" // anonymous import
 	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
 	"log"
 	http "net/http"
-	"reflect"
-	"runtime"
 )
 
 var (
@@ -107,77 +104,4 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-// APIResponse : Type used by the Response Channel
-// in the handlerWrapper (does not need json tags)
-type APIResponse struct {
-	Data       interface{}
-	StatusCode int
-	Error      error
-}
-
-func (a *APIResponse) setError(err error, errMessage string) {
-	a.setErrorAndStatus(http.StatusInternalServerError, err, errMessage)
-}
-
-func (a *APIResponse) setErrorAndStatus(status int, err error, errMessage string) {
-	a.StatusCode = status
-	a.Error = errors.Wrap(err, errMessage)
-}
-
-func (a *APIResponse) setResponse(data interface{}) {
-	a.setResponseAndStatus(http.StatusOK, data)
-}
-
-func (a *APIResponse) setResponseAndStatus(status int, data interface{}) {
-	a.StatusCode = status
-	a.Data = data
-}
-
-func (a *APIResponse) setStatus(status int) {
-	a.StatusCode = status
-}
-
-func handlerWrapper(handler func(r *http.Request, ar *APIResponse)) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		ar := APIResponse{nil, 200, nil}
-		funcName := runtime.FuncForPC(reflect.ValueOf(handler).Pointer()).Name()
-		log.Printf("New request:\n |url:  %s\n |func: %s\n", r.URL, funcName)
-		handler(r, &ar)
-
-		if ar.Error != nil {
-			log.Printf("Server error: %v", ar.Error)
-			if ar.StatusCode == http.StatusInternalServerError {
-				ar.Error = errors.New(http.StatusText(http.StatusInternalServerError))
-			}
-			http.Error(w, ar.Error.Error(), ar.StatusCode)
-			return
-		}
-
-		if ar.Data == nil {
-			w.WriteHeader(ar.StatusCode)
-			return
-		}
-
-		jsonData, err := json.Marshal(ar.Data)
-		if err != nil {
-			err := errors.Wrap(err, "Error during JSON Decoding")
-			log.Printf("Error marshalling response: %v", err)
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(ar.StatusCode)
-		_, err = w.Write(jsonData) //returns an integer, not sure what it's used for
-		if err != nil {
-			log.Printf("Error sending response to request: %v", err)
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
-		}
-
-		return
-	})
 }
